Change global history data structure
This commit changes the data structure of the global history from an
unsigned 64 bit shift register to an array. We are doing this because
we want an easy way to run benchmarks with very different global
history lengths. Having the array allows us to test very large history
lengths without changing the underlying code.